fix: cast SQL aggregates to bigint for Rust i64 compatibility
PostgreSQL SUM() and COUNT() return NUMERIC type, but Rust expects i64.
Added ::bigint casts to all aggregate queries.

+++ backend/src/handlers/folder.rs
@@ -82,7 +82,7 @@ pub async fn list_folders(
     for folder in folders {
         let stats = sqlx::query_as::<_, (Option<i64>, Option<i64>)>(
             r#"
-            SELECT COUNT(*), COALESCE(SUM(size), 0)
+            SELECT COUNT(*)::bigint, COALESCE(SUM(size), 0)::bigint
             FROM files
             WHERE folder_id = $1
             "#,
